fix(orch): serialize store mutations
