+++ plugins/pstack/skills/poteto-mode/scripts/orch/frontier.ts
@@ -1,0 +1,322 @@
+import { execFileSync } from "node:child_process";
+import {
+  UserError,
+  type Frontier,
+  type FrontierPr,
+  type FrontierPrState,
+} from "./types.ts";
+
+function errorMessage(error: unknown): string {
+  return error instanceof Error ? error.message : String(error);
+}
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return value !== null && typeof value === "object" && !Array.isArray(value);
+}
+
+function isUnknownArray(value: unknown): value is readonly unknown[] {
+  return Array.isArray(value);
+}
+
+function frontierPrStateOrNull(value: unknown): FrontierPrState | null {
+  switch (value) {
+    case "OPEN":
+    case "MERGED":
+    case "CLOSED":
+      return value;
+    default:
+      return null;
+  }
+}
+
+export function parseFrontier(raw: string): Frontier {
+  let value: unknown;
+  try {
+    value = JSON.parse(raw);
+  } catch {
+    throw new UserError("frontier.json is not valid JSON");
+  }
+  if (!isRecord(value)) {
+    throw new UserError("frontier.json must contain an object");
+  }
+  if (Object.keys(value).length === 0) {
+    return { generation: 0, prs: [], lowestUnmerged: null };
+  }
+  if (
+    typeof value.generation !== "number" ||
+    !Number.isSafeInteger(value.generation) ||
+    value.generation < 0 ||
+    !isUnknownArray(value.prs) ||
+    !(
+      value.lowestUnmerged === null ||
+      (typeof value.lowestUnmerged === "number" &&
+        Number.isSafeInteger(value.lowestUnmerged))
+    )
+  ) {
+    throw new UserError("frontier.json has an invalid shape");
+  }
+  const prs: FrontierPr[] = [];
+  for (const row of value.prs) {
+    const state = isRecord(row) ? frontierPrStateOrNull(row.state) : null;
+    if (
+      !isRecord(row) ||
+      typeof row.pr !== "number" ||
+      !Number.isSafeInteger(row.pr) ||
+      row.pr < 1 ||
+      typeof row.branches !== "string" ||
+      row.branches.length === 0 ||
+      typeof row.sha !== "string" ||
+      state === null
+    ) {
+      throw new UserError("frontier.json has an invalid PR row");
+    }
+    prs.push({
+      pr: row.pr,
+      branches: row.branches,
+      sha: row.sha,
+      state,
+    });
+  }
+  return {
+    generation: value.generation,
+    prs,
+    lowestUnmerged: value.lowestUnmerged,
+  };
+}
+
+const OPEN_GT_PR_STATUSES = new Set([
+  "Trunk branch locked",
+  "Changes requested",
+  "Waiting on PRs in this stack to merge",
+  "Waiting on downstack merge state",
+  "Draft",
+  "Required checks failed",
+  "Undergoing failure detection",
+  "Merge queue failed on current head commit",
+  "Handed off to merge queue...",
+  "Waiting on downstack",
+  "Merge conflicts",
+  "Needs reviewers",
+  "Needs approvals",
+  "Needs restack",
+  "Queued to merge...",
+  "Ready to merge",
+  "Ready to merge as stack",
+  "Rebasing...",
+  "Waiting on CI...",
+  "Stale, needs rebase onto trunk",
+  "Unresolved comments",
+  "Waiting on required CI",
+  "Waiting to merge...",
+]);
+
+interface GtPullRequest {
+  readonly pr: number;
+  readonly state: FrontierPrState;
+}
+
+interface GtFrontierEntry extends GtPullRequest {
+  readonly branches: string;
+}
+
+function parseGtPullRequest({
+  branch,
+  detail,
+}: {
+  branch: string;
+  detail: string;
+}): GtPullRequest {
+  const match =
+    /^(?:\[origin\] )?PR #([1-9]\d*)(?: \(([^)\r\n]+)\))?( .+)?$/.exec(
+      detail
+    );
+  const pr = Number(match?.[1] ?? 0);
+  if (match === null || !Number.isSafeInteger(pr)) {
+    throw new UserError(
+      `gt info output has an invalid PR row for branch ${branch}: ${detail}`
+    );
+  }
+  const status = match[2];
+  // A "(" right after the PR number that the status group did not capture
+  // means a nested-paren status like "Needs approvals (2)"; treating it as
+  // no-status would silently report the PR as OPEN.
+  if (status === undefined && (match[3] ?? "").startsWith(" (")) {
+    throw new UserError(
+      `gt info output has an invalid PR row for branch ${branch}: ${detail}`
+    );
+  }
+  if (status === "Merged") {
+    return { pr, state: "MERGED" };
+  }
+  if (status === "Closed") {
+    return { pr, state: "CLOSED" };
+  }
+  if (status === undefined || OPEN_GT_PR_STATUSES.has(status)) {
+    return { pr, state: "OPEN" };
+  }
+  throw new UserError(
+    `gt info output has an unknown PR state for branch ${branch}: ${status}`
+  );
+}
+
+function parseGtBranches(raw: string): readonly string[] {
+  const branches: string[] = [];
+  const lines = raw.replace(/\r/g, "").split("\n");
+  for (const [index, line] of lines.entries()) {
+    if (line.length === 0) {
+      continue;
+    }
+    // (?!-) rejects leading-dash names, which git and gt would parse as
+    // options when the branch is later passed to them as an argument.
+    const branchMatch =
+      /^(?:│ )*[◯◉] +((?!-)[^\s]+)((?: \([^()\r\n]*\))*)$/.exec(line);
+    if (branchMatch === null) {
+      throw new UserError(
+        `gt log short output has an unparseable line ${index + 1}: ${JSON.stringify(line)}`
+      );
+    }
+    const branch = branchMatch[1] ?? "";
+    if (branches.includes(branch)) {
+      throw new UserError(
+        `gt log short output contains duplicate branch ${branch}`
+      );
+    }
+    branches.push(branch);
+  }
+  const trunk = branches[0];
+  if (trunk === undefined) {
+    throw new UserError("gt log short output did not contain a stack");
+  }
+  return branches.slice(1);
+}
+
+function graphitePullRequest({
+  branch,
+  repo,
+}: {
+  branch: string;
+  repo: string;
+}): GtPullRequest {
+  let raw: string;
+  try {
+    raw = execFileSync("gt", ["--no-interactive", "info", branch], {
+      cwd: repo,
+      encoding: "utf8",
+      env: { ...process.env, NO_COLOR: "1" },
+      stdio: ["ignore", "pipe", "pipe"],
+    });
+  } catch (error) {
+    throw new UserError(`gt info ${branch} failed: ${errorMessage(error)}`);
+  }
+  const rows = raw
+    .replace(/\r/g, "")
+    .split("\n")
+    .filter(
+      (line) => line.startsWith("PR #") || line.startsWith("[origin] PR #")
+    );
+  if (rows.length === 0) {
+    throw new UserError(
+      `gt info output branch ${branch} has no pull request; this clone's gt metadata may predate the submit, so resolve the frontier from the stacker's clone or after gt sync`
+    );
+  }
+  if (rows.length > 1) {
+    throw new UserError(
+      `gt info output contains multiple PRs for branch ${branch}`
+    );
+  }
+  return parseGtPullRequest({ branch, detail: rows[0] ?? "" });
+}
+
+function graphiteFrontier(repo: string): readonly GtFrontierEntry[] {
+  let raw: string;
+  try {
+    raw = execFileSync(
+      "gt",
+      ["--no-interactive", "log", "short", "--stack", "--reverse"],
+      {
+        cwd: repo,
+        encoding: "utf8",
+        env: { ...process.env, NO_COLOR: "1" },
+        stdio: ["ignore", "pipe", "pipe"],
+      }
+    );
+  } catch (error) {
+    throw new UserError(
+      `gt log short --stack --reverse failed: ${errorMessage(error)}`
+    );
+  }
+  const result = parseGtBranches(raw).map((branch) => ({
+    branches: branch,
+    ...graphitePullRequest({ branch, repo }),
+  }));
+  if (new Set(result.map((row) => row.pr)).size !== result.length) {
+    throw new UserError("gt info output contains duplicate pull requests");
+  }
+  return result;
+}
+
+function branchSha({
+  branch,
+  repo,
+}: {
+  branch: string;
+  repo: string;
+}): string {
+  let raw: string;
+  try {
+    raw = execFileSync("git", ["rev-parse", branch], {
+      cwd: repo,
+      encoding: "utf8",
+      env: process.env,
+      stdio: ["ignore", "pipe", "pipe"],
+    });
+  } catch (error) {
+    throw new UserError(
+      `git rev-parse ${branch} failed: ${errorMessage(error)}`
+    );
+  }
+  const sha = raw.trim();
+  if (!/^[0-9a-f]{40,64}$/i.test(sha)) {
+    throw new UserError(`git rev-parse ${branch} returned an invalid SHA`);
+  }
+  return sha;
+}
+
+export function resolveFrontier(repo: string): readonly FrontierPr[] {
+  return graphiteFrontier(repo).map((row) => ({
+    ...row,
+    sha: branchSha({ branch: row.branches, repo }),
+  }));
+}
+
+export function validateFrontierPin({
+  actual,
+  expected,
+}: {
+  actual: readonly number[];
+  expected: readonly number[];
+}): void {
+  if (
+    actual.length === expected.length &&
+    actual.every((pr, index) => pr === expected[index])
+  ) {
+    return;
+  }
+  const actualSet = new Set(actual);
+  const expectedSet = new Set(expected);
+  const missing = expected.filter((pr) => !actualSet.has(pr));
+  const extra = actual.filter((pr) => !expectedSet.has(pr));
+  const drift: string[] = [];
+  if (missing.length > 0) {
+    drift.push(`missing from gt: ${missing.join(",")}`);
+  }
+  if (extra.length > 0) {
+    drift.push(`extra in gt: ${extra.join(",")}`);
+  }
+  if (missing.length === 0 && extra.length === 0) {
+    drift.push(
+      `order differs: expected ${expected.join(",")}; gt ${actual.join(",")}`
+    );
+  }
+  throw new UserError(`frontier pin mismatch: ${drift.join("; ")}`);
+}
