+++ plugins/pstack/skills/poteto-mode/scripts/orch/status.ts
@@ -1,0 +1,254 @@
+import type {
+  Counts,
+  Frontier,
+  Gate,
+  LedgerEntry,
+  OpenGate,
+  StatusSummary,
+  Unit,
+} from "./types.ts";
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return value !== null && typeof value === "object" && !Array.isArray(value);
+}
+
+function isUnknownArray(value: unknown): value is readonly unknown[] {
+  return Array.isArray(value);
+}
+
+export function countValues(values: readonly string[]): Counts {
+  const result: Record<string, number> = {};
+  for (const value of values) {
+    result[value] = (result[value] ?? 0) + 1;
+  }
+  return Object.fromEntries(
+    Object.entries(result).sort(([left], [right]) => left.localeCompare(right))
+  );
+}
+
+export function summarize(
+  unitRows: readonly Unit[],
+  ledgerRows: readonly LedgerEntry[],
+  currentFrontier: Frontier,
+  gateRows: readonly Gate[]
+): StatusSummary {
+  return {
+    unitStates: countValues(unitRows.map((unit) => unit.state)),
+    ledgerVerdicts: countValues(ledgerRows.map((row) => row.verdict)),
+    frontierGeneration: currentFrontier.generation,
+    openGateIds: gateRows
+      .filter((gate): gate is OpenGate => gate.kind === "open")
+      .map((gate) => gate.id)
+      .sort(),
+  };
+}
+
+function countRecord(value: unknown): Record<string, number> | null {
+  if (!isRecord(value)) {
+    return null;
+  }
+  const result: Record<string, number> = {};
+  for (const [name, count] of Object.entries(value)) {
+    if (
+      typeof count !== "number" ||
+      !Number.isSafeInteger(count) ||
+      count < 0
+    ) {
+      return null;
+    }
+    result[name] = count;
+  }
+  return result;
+}
+
+export function previousSummary(raw: string): StatusSummary | null {
+  const match = /<!-- orch-summary (.+) -->/.exec(raw);
+  if (match === null) {
+    return null;
+  }
+  let value: unknown;
+  try {
+    value = JSON.parse(match[1] ?? "");
+  } catch {
+    return null;
+  }
+  if (
+    !isRecord(value) ||
+    typeof value.frontierGeneration !== "number" ||
+    !isUnknownArray(value.openGateIds)
+  ) {
+    return null;
+  }
+  const unitStates = countRecord(value.unitStates);
+  const ledgerVerdicts = countRecord(value.ledgerVerdicts);
+  const openGateIds = value.openGateIds.filter(
+    (item): item is string => typeof item === "string"
+  );
+  if (
+    unitStates === null ||
+    ledgerVerdicts === null ||
+    openGateIds.length !== value.openGateIds.length
+  ) {
+    return null;
+  }
+  return {
+    unitStates,
+    ledgerVerdicts,
+    frontierGeneration: value.frontierGeneration,
+    openGateIds,
+  };
+}
+
+export function changed(
+  before: StatusSummary | null,
+  after: StatusSummary
+): string {
+  if (before === null) {
+    return "first render";
+  }
+  const result: string[] = [];
+  const groups: readonly {
+    readonly label: string;
+    readonly oldCounts: Counts;
+    readonly newCounts: Counts;
+  }[] = [
+    {
+      label: "units",
+      oldCounts: before.unitStates,
+      newCounts: after.unitStates,
+    },
+    {
+      label: "ledger",
+      oldCounts: before.ledgerVerdicts,
+      newCounts: after.ledgerVerdicts,
+    },
+  ];
+  for (const { label, oldCounts, newCounts } of groups) {
+    const names = [
+      ...new Set([...Object.keys(oldCounts), ...Object.keys(newCounts)]),
+    ].sort();
+    for (const name of names) {
+      const oldCount = oldCounts[name] ?? 0;
+      const newCount = newCounts[name] ?? 0;
+      if (oldCount !== newCount) {
+        result.push(`${label} ${name} ${oldCount}->${newCount}`);
+      }
+    }
+  }
+  if (before.frontierGeneration !== after.frontierGeneration) {
+    result.push(
+      `frontier generation ${before.frontierGeneration}->${after.frontierGeneration}`
+    );
+  }
+  if (before.openGateIds.join("\0") !== after.openGateIds.join("\0")) {
+    result.push(
+      `open gates ${before.openGateIds.length}->${after.openGateIds.length}`
+    );
+  }
+  return result.length === 0 ? "no derived changes" : result.join("; ");
+}
+
+function markdown(value: string): string {
+  return value
+    .replace(/[\t\n\r]/g, " ")
+    .replace(/\\/g, "\\\\")
+    .replace(/\|/g, "\\|");
+}
+
+function table(
+  headers: readonly string[],
+  rows: readonly (readonly string[])[]
+): string {
+  if (rows.length === 0) {
+    return "(none)";
+  }
+  return [
+    `| ${headers.join(" | ")} |`,
+    `| ${headers.map(() => "---").join(" | ")} |`,
+    ...rows.map((row) => `| ${row.map(markdown).join(" | ")} |`),
+  ].join("\n");
+}
+
+function countLine(value: Counts): string {
+  const entries = Object.entries(value);
+  return entries.length === 0
+    ? "none"
+    : entries.map(([name, count]) => `${name}=${count}`).join(", ");
+}
+
+export function statusMarkdown(
+  unitRows: readonly Unit[],
+  ledgerRows: readonly LedgerEntry[],
+  currentFrontier: Frontier,
+  gateRows: readonly Gate[],
+  currentSummary: StatusSummary
+): string {
+  return `# Orchestrate status
+
+Generated: ${new Date().toISOString()}
+
+## Units
+
+States: ${countLine(currentSummary.unitStates)}
+
+${table(
+  ["ID", "Track", "State", "Branch", "PR", "SHA", "Brief"],
+  unitRows.map((unit) => [
+    unit.id,
+    unit.track,
+    unit.state,
+    unit.branch,
+    unit.pr,
+    unit.sha,
+    unit.brief,
+  ])
+)}
+
+## Verification ledger
+
+Verdicts: ${countLine(currentSummary.ledgerVerdicts)}
+
+${table(
+  ["PR", "SHA", "Verdict", "Evidence", "Verifier", "Timestamp"],
+  ledgerRows.map((row) => [
+    row.pr,
+    row.sha,
+    row.verdict,
+    row.evidence,
+    row.verifier,
+    row.ts,
+  ])
+)}
+
+## Frontier
+
+Generation: ${currentFrontier.generation}
+Lowest unmerged: ${currentFrontier.lowestUnmerged ?? "none"}
+
+${table(
+  ["Branch", "PR", "SHA", "State"],
+  currentFrontier.prs.map((row) => [
+    row.branches,
+    String(row.pr),
+    row.sha,
+    row.state,
+  ])
+)}
+
+## Gates
+
+${table(
+  ["ID", "Status", "Question", "Options", "Default", "Answer"],
+  gateRows.map((gate) => [
+    gate.id,
+    gate.kind,
+    gate.question,
+    gate.options,
+    gate.defaultAnswer,
+    gate.kind === "resolved" ? gate.answer : "",
+  ])
+)}
+
+<!-- orch-summary ${JSON.stringify(currentSummary)} -->
+`;
+}
